Cre8Hub-AI-Workflow/requirements.txt
fastapi
uvicorn[standard]

# LangChain & related components (pinned to 0.2.x for LLMChain compatibility)
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langchain-huggingface

# Embeddings
sentence-transformers
huggingface-hub

# Vector store
faiss-cpu

# Typing & Logging (built-in, no need to install separately)
# typing
# logging

# If you're using dotenv to load env vars
python-dotenv
pydantic
# Google GenAI (compatible with both cre8echo and cre8canvas)
langchain-google-genai==1.0.10
google-generativeai>=0.8.0
pandas
python-multipart
qdrant-client
pytrends
logger
requests
redis
pymongo

# Image processing
Pillow>=10.0.0